pyproject.toml
[tool.poetry]
name = "dispair"
version = "0.0.1"
description = "Webhook handler for Discord Interactions"
authors = ["GDWR <gregory.dwr@gmail.com>"]
homepage = "https://github.com/GDWR/Dispair"
repository = "https://github.com/GDWR/Dispair"
readme = "README.md"
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.9",
    "Typing :: Typed",
]

[tool.poetry.dependencies]
python = "^3.9"
PyNaCl = "^1.4.0"
aiohttp = "^3.7.4"
flake8 = "^3.9.1"

[tool.poetry.dev-dependencies]
flake8 = "~=3.8"
flake8-annotations = "~=2.3"
flake8-bugbear = "~=20.1"
flake8-docstrings = "~=1.5"
flake8-import-order = "~=0.18"
flake8-string-format = "~=0.3"
flake8-tidy-imports = "~=4.1"
flake8-todo = "~=0.7"
pep8-naming = "~=0.11"
pre-commit = "~=2.1"
taskipy = "^1.6.0"
pytest = "^5.2"
pytest-asyncio = "^0.15.1"
pytest-timeout = "^1.4.2"
python-dotenv = "^0.17.1"

[tool.poetry.scripts]
example = 'example.__main__:main'

[tool.poetry.urls]
"Bug Tracker" = "https://github.com/GDWR/Dispair/issues"

[tool.taskipy.tasks]
lint = "pre-commit run --all-files"
precommit = "pre-commit install"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
